+++ src/message/encode.js
@@ -3,9 +3,6 @@
 const binary = require('binary-data');
 const { StunMessagePacket, StunAttributePacket } = require('../lib/protocol');
 
-const kCookie = Symbol.for('kCookie');
-const kAttributes = Symbol.for('kAttributes');
-
 const {
   types: { array },
 } = binary;
@@ -13,15 +10,14 @@ const {
 module.exports = encode;
 
 /**
- * Encode the StunMessage to the Buffer.
+ * Encode the StunMessage to a Buffer.
  * @param {StunMessage} message
  * @returns {Buffer}
  */
 function encode(message) {
   const ostream = binary.createEncode();
-  const attrlist = message[kAttributes];
 
-  const attributes = attrlist.map((attribute) => ({
+  const attributes = message._attributes.map((attribute) => ({
     type: attribute.type,
     value: attribute.toBuffer(),
   }));
@@ -33,7 +29,7 @@ function encode(message) {
         attributes,
         array(StunAttributePacket, attributes.length),
       ),
-      cookie: message[kCookie],
+      cookie: message._cookie,
       transaction: message.transactionId,
     },
     attributes,